#52: implemented two specific obs sink plugins for simple and download types
