+++ src/org/aavso/tools/vstar/ui/resources/locale/strings.properties
@@ -269,6 +269,8 @@ PERIOD_ANALYSIS=Period Analysis
 // Plug-in descriptions
 DATABASE_OBS_SOURCE=AAVSO International Database observation source
 TEXT_FORMAT_FILE=Download or Simple
+DOWNLOAD_FORMAT_FILE=Download
+SIMPLE_FORMAT_FILE=Simple
 
 // Other
 THE=The
