+++ src/org/aavso/tools/vstar/ui/resources/locale/strings_es.properties
@@ -276,6 +276,8 @@ PERIOD_ANALYSIS=An\u00E1lisis del Per\u00EDodo
 // Plug-in descriptions
 DATABASE_OBS_SOURCE=Fuente de observaciones AAVSO International Database
 TEXT_FORMAT_FILE=Descarga o simple
+DOWNLOAD_FORMAT_FILE=Descarga
+SIMPLE_FORMAT_FILE=Simple
 
 // Other
 THE=El
