+++ src/org/aavso/tools/vstar/ui/resources/locale/strings_fr.properties
@@ -276,6 +276,8 @@ PERIOD_ANALYSIS=analyse de la p\u00E9riode
 // Plug-in descriptions
 DATABASE_OBS_SOURCE=AAVSO International Database (AID), la r\u00E9f\u00E9rence pour les donn\u00E9es d'observation
 TEXT_FORMAT_FILE=T\u00E9l\u00E9charger ou simple
+DOWNLOAD_FORMAT_FILE=T\u00E9l\u00E9charger
+SIMPLE_FORMAT_FILE=Simple
 
 // Other
 THE=le
